fix: prevent helm release timeouts
Add a workaround to prevent Helm release timeouts by changing the
deployment strategy of the Helm chart.

Rolling updates of single replica deployments in combination with
ReadWriteOnce PVCs lead to a scenario where the new pod will never start
successfully since it fails to attach to the PVC that is still used by
the old pod. The recreate strategy kills the old pod before starting the
new one.